requirements.txt
aiohttp
motor
pyrogram==2.0.106
pyromod
pytz
python-dotenv
pykeyboard
attrify
pillow
qrcode[pil]
gpytranslate
pymediainfo
requests
tornado
psutil
beautifulsoup4
python-dateutil
telegraph
filetype
wget
gtts
yt_dlp
